pok-08: refactor: utilize interfaces instead types
