+++ src/app/pages/favorites/favorites.page.ts
@@ -23,7 +23,7 @@ import {
 import { FavButtonComponent } from '../../components/fav-button/fav-button.component';
 import { ToolbarComponent } from '../../components/toolbar/toolbar.component';
 import { Router } from '@angular/router';
-import { SimplePokemon } from '../../types/Pokemon';
+import { SimplePokemon } from '../../interfaces/pokemons.interface';
 import { FavoriteService } from '../../services/favorite.service';
 import { SearchService } from '../../services/search.service';
 import { addIcons } from 'ionicons';
